chore: Add option to remove inactive users from seat assignments
This commit adds a new input option to the action.yml file, allowing users to remove only the inactive users from seat assignments. The option is called "inactive_only" and is set to false by default. Additionally, a new input option called "inactive_days" is added to specify the number of days to identify an inactive user, with a default value of 25.

Note: This commit message has been generated based on the provided code changes and recent repository commits.